+++ Slack.NetStandard/Slack.NetStandard.csproj
@@ -13,7 +13,7 @@
     <RepositoryUrl>https://github.com/stoiveyp/Slack.NetStandard</RepositoryUrl>
     <PackageTags>.net, slack</PackageTags>
     <LangVersion>latest</LangVersion>
-    <AssemblyVersion>1.0.0.0</AssemblyVersion>
+    <AssemblyVersion>1.1.0.0</AssemblyVersion>
   </PropertyGroup>
 
   <ItemGroup>
